created models for the frontend
