+++ djv-web/src/main/webapp/app/api/user.ts
@@ -1,19 +1,7 @@
-import {fetchData, HttpMethod} from "./utils";
+import { IUser } from "../model/User";
+import { fetchData, HttpMethod } from "./utils";
 
 const PATH_PREFIX = "/api/user";
-
-enum UserType {
-    REGULAR,
-    ADMIN
-}
-
-interface IUser {
-    firstName?: string;
-    lastName?: string;
-    email: string;
-    type: UserType;
-    banned: boolean;
-}
 
 export const getUsers = (): Promise<IUser[]> => fetchData(PATH_PREFIX + "/", HttpMethod.GET);
