+++ djv-web/src/main/webapp/app/model/Category.ts
@@ -1,0 +1,6 @@
+export interface ICategory {
+    name: string;
+    icon: string;
+    displayName: string;
+    children: ICategory[];
+}
