+++ djv-web/src/main/webapp/app/model/User.ts
@@ -1,0 +1,13 @@
+// api/user.ts yra user type ir a user. reik situos issikelt
+enum UserType {
+    REGULAR,
+    ADMIN
+}
+
+export interface IUser {
+    firstName?: string;
+    lastName?: string;
+    email: string;
+    type: UserType;
+    banned: boolean;
+}
